- Moving class definition in header file, needed by DOXYGEN for documentation.  - Fix using only the first sample for the recHit definition
